+++ common/log/src/lib.rs
@@ -20,9 +20,12 @@
 )]
 
 mod capture_logs;
-
 pub use capture_logs::{ChildCapture, ParsedMsg};
-use opentelemetry::sdk::propagation::TraceContextPropagator;
+use opentelemetry::{
+    sdk::{propagation::TraceContextPropagator, trace, Resource},
+    KeyValue,
+};
+use opentelemetry_otlp::WithExportConfig;
 use std::net::IpAddr;
 use tracing::level_filters::LevelFilter;
 pub use tracing::{self, Level, Span};
@@ -35,10 +38,14 @@ use tracing_subscriber::{fmt, layer::SubscriberExt, Layer, Registry};
 /// Allow to configure the tracing.
 #[derive(Clone, Debug)]
 pub struct TelemetryConfig {
-    /// Endpoint of Jaeger server to send logs to.
-    pub jaeger_endpoint: Option<String>,
-    /// Service name to collect logs on Jaeger.
-    pub jaeger_service_name: Option<String>,
+    /// Endpoint of [Opentelemetry] collector server to send logs to.
+    ///
+    /// [Opentelemetry]: https://opentelemetry.io
+    pub otlp_endpoint: Option<String>,
+    /// Service name to collect traces to [Opentelemetry] collector.
+    ///
+    /// [Opentelemetry]: https://opentelemetry.io
+    pub service_name: Option<String>,
     /// Logging level
     pub level: LevelFilter,
 }
@@ -52,54 +59,69 @@ impl TelemetryConfig {
     pub fn new(level: Option<Level>) -> Self {
         Self {
             level: LevelFilter::from(level.unwrap_or(Level::INFO)),
-            jaeger_endpoint: None,
-            jaeger_service_name: None,
+            otlp_endpoint: None,
+            service_name: None,
         }
     }
 
-    /// Set Jaeger endpoint to send traces to.
+    /// Set [Opentelemetry] collector endpoint to send traces to.
+    ///
+    /// [Opentelemetry]: https://opentelemetry.io
     #[must_use]
-    pub fn jaeger_endpoint(
+    pub fn otlp_endpoint(
         mut self,
-        agent_ip: Option<IpAddr>,
-        agent_port: Option<u16>,
+        collector_ip: Option<IpAddr>,
+        collector_port: Option<u16>,
     ) -> Self {
-        if let (Some(ip), Some(port)) = (agent_ip, agent_port) {
-            self.jaeger_endpoint = Some(format!("{ip}:{port}"));
+        if let (Some(ip), Some(port)) = (collector_ip, collector_port) {
+            self.otlp_endpoint = Some(format!("http://{ip}:{port}"));
         };
         self
     }
 
-    /// Set Jaeger service name to collect logs on Jaeger.
+    /// Set current service name to collect traces to [Opentelemetry] collector.
+    ///
+    /// [Opentelemetry]: https://opentelemetry.io
     #[must_use]
-    pub fn jaeger_service_name(mut self, service_name: String) -> Self {
-        self.jaeger_service_name = Some(service_name);
+    pub fn service_name(mut self, service_name: String) -> Self {
+        self.service_name = Some(service_name);
         self
     }
 
     /// Initialize the logging and telemetry.
-    /// If [`TelemetryConfig.jaeger_endpoint`] is set,
-    /// the telemetry will be sent to Jaeger
+    /// If [`TelemetryConfig.otlp_endpoint`] is set,
+    /// the telemetry will be sent to [Opentelemetry] collector.
     ///
     /// # Panics
     ///
     /// If failed to initialize logger.
+    ///
+    /// [Opentelemetry]: https://opentelemetry.io
     pub fn init(self) {
         if let Err(e) = LogTracer::init() {
             panic!("Failed to initialize logger: {e}");
         };
-        let service_name = self.jaeger_service_name.unwrap_or("unknown".into());
+        let service_name = self.service_name.unwrap_or("unknown".into());
 
-        let mut layers = vec![fmt::layer().pretty().boxed()];
+        let mut layers = vec![fmt::layer().compact().boxed()];
 
-        if let Some(endpoint) = self.jaeger_endpoint {
-            let tracer = opentelemetry_jaeger::new_agent_pipeline()
-                .with_endpoint(endpoint)
-                .with_service_name(service_name)
-                .with_max_packet_size(9216)
-                .with_auto_split_batch(true)
-                .install_batch(opentelemetry::runtime::Tokio)
-                .expect("Failed to install jaeger agent");
+        if let Some(endpoint) = self.otlp_endpoint {
+            let otlp_exporter = opentelemetry_otlp::new_exporter()
+                .tonic()
+                .with_endpoint(endpoint);
+
+            let trace_config =
+                trace::config().with_resource(Resource::new(vec![
+                    KeyValue::new("service.name", service_name),
+                ]));
+
+            let tracer = opentelemetry_otlp::new_pipeline()
+                .tracing()
+                .with_exporter(otlp_exporter)
+                .with_trace_config(trace_config)
+                .install_simple()
+                .expect("Failed to install OTLP tracer");
+
             opentelemetry::global::set_text_map_propagator(
                 TraceContextPropagator::new(),
             );