+++ components/restreamer/src/server.rs
@@ -29,8 +29,8 @@ use crate::{
 #[actix_web::main]
 pub async fn run(mut cfg: Opts) -> Result<(), Failure> {
     TelemetryConfig::new(cfg.verbose)
-        .jaeger_endpoint(cfg.jaeger_agent_ip, cfg.jaeger_agent_port)
-        .jaeger_service_name(cfg.jaeger_service_name.clone())
+        .otlp_endpoint(cfg.otlp_collector_ip, cfg.otlp_collector_port)
+        .service_name(cfg.service_name.clone())
         .init();
 
     if cfg.public_host.is_none() {